Refactor for packaging.
